init file in feature_extractor emptied
everything moved to features module, this way it's easier to extend the list of features.
